pull out most of start task command to be separate from discord
