+++ commands/db/stop-timer.py
@@ -7,8 +7,8 @@ from dotenv import load_dotenv
 from sqlalchemy import update, select
 
 from api.db_classes import SpeedTask, get_session, Submissions
-from api.task_handling import is_time_over
-from api.utils import is_task_currently_running, get_host_role, get_submitter_role
+from api.task_handling import is_time_over, is_task_currently_running
+from api.utils import get_host_role, get_submitter_role
 from commands.db.requesttask import has_requested_already
 
 load_dotenv()
