+++ events/on_ready.py
@@ -1,16 +1,18 @@
-import discord
-import shared
-import time
 import asyncio
 import os
-from dotenv import load_dotenv
-from discord.ext import commands, tasks
+import time
 
-from api.submissions import get_logs_channel
-from api.utils import is_task_currently_running, get_tasks_channel, get_announcement_channel, get_submitter_role
-from api.db_classes import get_session, Tasks, SpeedTaskDesc, SpeedTaskLength, ReminderPings, SpeedTaskReminders, \
-    SpeedTask, Submissions
+import discord
+import shared
+from discord.ext import commands, tasks
+from dotenv import load_dotenv
 from sqlalchemy import select, update, delete, insert
+
+from api.db_classes import get_session, Tasks, SpeedTaskDesc, SpeedTaskLength, ReminderPings, SpeedTaskReminders, \
+    SpeedTask
+from api.submissions import get_logs_channel
+from api.task_handling import is_task_currently_running
+from api.utils import get_tasks_channel, get_announcement_channel, get_submitter_role
 
 load_dotenv()
 DEFAULT = os.getenv('DEFAULT')  # Choices: mkw, sm64
